requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.14.0
python-dotenv==0.21.0
requests==2.28.1
celery==5.2.7
requests-futures==1.0.0
django-cprofile-middleware==1.0.5